% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIPS.COMID.R
\name{map.FIPS}
\alias{map.FIPS}
\title{COMID and FIPS map}
\usage{
visualize where COMID is
}
\arguments{
\item{overlay}{as sf object}
}
\value{
map with points that represent NHDPlus coordinates for that given COMID
}
\description{
after running 1 of the 3 COMID to FIPS functions, simply enter "overlay" as the sole argument to return the map of COMID within its county code
}
\examples{
map.FIPS(overlay)
}
\seealso{
[COMID.FIPS()][map.FIPSflow()][map.FIPSzoomout()]
}
\author{
Kent Codding
}
