% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIPS.COMID.R
\name{map.FIPSzoomout}
\alias{map.FIPSzoomout}
\title{zoomed out COMID and FIPS map}
\usage{
map.FIPSzoomout(overlay)
}
\description{
this function is just a zoomed out version of map.FIPS
}
\author{
Kent Codding
}
