% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIPS.COMID.R
\name{map.FIPSflow}
\alias{map.FIPSflow}
\title{COMID and FIPS flowline map}
\usage{
visualize where COMID is
}
\arguments{
\item{overlay}{as sf object}
}
\value{
zoomed in map of county codes that represent NHDPlus coordinates for that given COMID. COMIDs with larger spatial extent typically show a flowline in the form of FIPS codes
}
\description{
after running 1 of the 3 COMID to FIPS functions, simply enter "overlay" as the sole argument to return the map of COMID within its county code
}
\examples{
map.FIPS(overlay)
}
\author{
Kent Codding
}
